// src/types/mock.ts
export interface DicomReport {
  findings: string[];
  impression: string[];
  hasAbnormal: boolean;
}

export interface DicomSeries {
  id: string;
  description: string;
  path: string;
}

export interface DicomStudy {
  id: string;
  type: string;
  date: string;
  description: string;
  path: string;
  report: DicomReport;
  series: DicomSeries[];
}

export interface DicomStudies {
  studies: DicomStudy[];
}